requirements.txt
requests
selectorlib
pypdf
google-genai
python-dotenv